Created Scene for Playtest 3
Quite a lot of changes. Overall cleaned up the project. Changed gun bullet value to 3. Changed enemy health to 4.

Created an entirely new scene and put old scenes in a folder to avoid clutter. Also removed the gun from the player prefab so we can start the game with no items.